requirements.txt
pillow==10.0.1
numpy==1.26.0
